make sd card script: Remove some hardcoded dependencies
This removed the dependency on the global layout of the project directory. Some projects put seL4
tools in the tools directory and some put it in the projects directory. This now uses a relative
path to find the make sd card script. This also uses the elements of the cmake project to determine
the name of the executable to copy to the sd card.