Change dependencies of the boot environment configuration
This moves the boot environment configuration to the end of the rootserver cmake file, allowing it
to depend on the creation of the final elfloader/seL4/application binary. This allows creating an
HSS payload that depends on the final seL4 build image as opposed to just supporting uboot, which is
what the previous version only supported.