Change how the Bin2Chunk Payload is set
Before it would be automatically set to the uboot binary by default. Now it is set to default to
FALSE because then simple conditionals can be used to set the payload to be the final sel4 image
name later. This must be done later and can not be done at the first pass when variables are
calculated because the image name is built up in the cmake function that declares the rootserver.