Use CMake lists instead of strings for configuring boot env
This required parsing the arguments to the ConfigureBootEnv with the cmake_parse_arguments function
because otherwise if a list is provided to the function, each element of the list becomes it's own
individual argument. So the cmake_parse_arguments function is used to keep them all in a list, which
makes it easier to use the IN_LIST functionality of the cmake "if" command when determining which
boot environments to build for.